Move starboard logic to data::starboard
