+++ src/event_handlers/channel_delete.rs
@@ -1,4 +1,4 @@
-use crate::{serenity, Data, Error, commands::starboard::delete_starboard_tables};
+use crate::{serenity, Data, Error, data::starboard::delete_starboard_tables};
 use serenity::GuildChannel;
 
 pub async fn handle(deleted_channel: &GuildChannel, data: &Data) -> Result<(), Error> {
