+++ src/event_handlers/reaction_add.rs
@@ -1,7 +1,5 @@
-use crate::{serenity, Data, Error};
+use crate::{serenity, Data, Error, data::starboard::manage_starboard_entry};
 use serenity::{Context, Reaction};
-
-use super::starboard_handler::manage_starboard_entry;
 
 pub async fn handle(reaction: &Reaction, data: &Data, ctx: &Context) -> Result<(), Error> {
     manage_starboard_entry(ctx, data, reaction).await?;
